Changed the type of variables from 'int' to 'float'

+++ Unit_2_C_Programming/Lesson_3_C_Basics/Assignments/Homework_1/EX_7/ex7.c
@@ -13,16 +13,16 @@
 
 int main(int argc, char** argv)
 {
-    int a = 0;
-    int b = 0;
+    float a = 0;
+    float b = 0;
     
     printf("Please enter the value of a: ");
     fflush(stdin); fflush(stdout);
-    scanf("%d", &a);
+    scanf("%f", &a);
 
     printf("Please enter the value of b: ");
     fflush(stdin); fflush(stdout);
-    scanf("%d", &b);
+    scanf("%f", &b);
 
     /*
     ex.
@@ -34,6 +34,6 @@ int main(int argc, char** argv)
     b = a / b;
     a /= b;
     
-    printf("The new value of a = %d\nThe new value of b = %d\n", a, b);
+    printf("The new value of a = %f\nThe new value of b = %f\n", a, b);
     return 0;
 }
